[Modify] Update LP Pool Card for frontend
